optimization v2 fix1

+++ app/core/model_manager.py
@@ -32,7 +32,7 @@ class ModelManager:
             # 4-bit Quantization Configuration for T4 Speedup
             quant_config = BitsAndBytesConfig(
                 load_in_4bit=True,
-                bnb_4bit_compute_dtype=torch.bfloat16,
+                bnb_4bit_compute_dtype=torch.float16, # float16 is safer/native for T4
                 bnb_4bit_use_double_quant=True,
                 bnb_4bit_quant_type="nf4"
             )
